requirements.txt
aiogram==3.2.0
aiohttp==3.8.1
python-dotenv==1.0.0
